backend/requirements.txt
flask>=3.0
firebase-admin>=6.0
python-dotenv>=1.0
flask-cors>=4.0
openai
gunicorn>=21.0
